+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-01-31 03:54:08
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 3dfde90bd508b46a572efbad73fa2d49b0250a9b
+
+## [ForwardWidgets-2kuai] - 2026-01-31 03:54:08
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 67b087d2704d0c6b7a437b1fe19fd6a3623ccdc9
+
+## [ForwardWidgets-pack1r] - 2026-01-31 03:54:08
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 7eb4f8eb0cb3dc0fc3cd583007adcb548f500096
+
+
 ## [ForwardWidgets-huangxd] - 2026-01-30 03:56:44
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main